RAG_System/Vision/requirements.txt
# Core Data Science (compatible con Python 3.9)
numpy>=1.21.0,<2.0.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
joblib>=1.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0

# Recommender Systems
implicit>=0.6.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.5.0
albumentations>=1.0.0

# Datasets
datasets>=2.0.0
huggingface-hub>=0.10.0

# LLM & AI APIs
openai>=1.0.0

# Application
streamlit>=1.20.0
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
python-multipart>=0.0.5

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.60.0
python-dotenv>=0.19.0
pyyaml>=6.0
requests>=2.27.0
pyarrow>=8.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code Quality (opcional)
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
